requirements.txt
flask
flask-sqlalchemy
flask-security-too
flask-mailman
bcrypt
python-dotenv